requirements.txt
streamlit
pandas
plotly
python-dotenv
MetaTrader5
requests